mysql fix, and registration fix

+++ database/migrations/2021_05_13_103136_create_sold__products_table.php
@@ -15,9 +15,9 @@ class CreateSoldProductsTable extends Migration
     {
         Schema::create('sold__products', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products');
             $table->timestamps();
         });
